refactor: move createTextLikeController out of public API
The createTextLikeController helper and its createStringListInputProps
utility are internal to useFilterBuilder's value input construction.
Move them from types.ts to useFilterBuilder.ts so they are no longer
exported from the package.
